asdsadas

+++ src/utils/string_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   string_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fborroto <fborroto@student.42.fr>          +#+  +:+       +#+        */
+/*   By: evocatur <evocatur@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/24 22:21:18 by edoardo           #+#    #+#             */
-/*   Updated: 2024/04/18 18:21:58 by fborroto         ###   ########.fr       */
+/*   Updated: 2024/04/26 17:28:35 by evocatur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,9 +35,9 @@ int	check_extension(char *file)
 	return (0);
 }
 
-size_t	last_idx(char *map_line)
+int	last_idx(char *map_line)
 {
-	size_t	i;
+	int	i;
 
 	i = ft_strlen(map_line) - 1;
 	while (is_spaces(map_line[i]))
